Bot-Development/Discord-Telegram-Bots/level-3/todobot_AdyaArchita/requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
